Initial implementation of taint propagation
Each TValue and headered object now has a "taint" field which is a
pointer to a constant struct of a new type named lua_TaintInfo.

The core defines of the VM will now, when manipulating and copying
objects, propagate the taint field on each value to the destination
and will update the global taint status of the Lua state.

An additional manipulation is required to ensure that all values
passed as arguments on the stack in luaD_precall are additionally
tainted if the state is currently insecure, and to also set the state
to insecure if the executed closure has a taint flag set in its header.

One major change here is that some macros for assigning simple values
like booleans, nils, and numbers to TValues now need to pass the Lua
state as their initial parameter.