Expose Lua APIs for taint querying and control
This exposes the WoW Lua functions such as forceinsecure, issecure,
issecurevariable, and securecall to the global environment allowing
them to query the taint associated with the global state or any
accessible table field, for secure code to invoke insecure code
with a taint barrier to prevent propagation back to the caller.

The geterrorhandler and seterrorhandler pair are also present; the
global error handler set is invoked if a securecall fails with an error.

An additional "debug.forcesecure" function is exposed which sets the
taint status of the Lua state back to a secure one, which can be
useful for debugging.