Set error handler in test script
Had accidentally removed this prior which meant that any errors
raised in securecall invokations weren't being displayed - however
there were no errors so this was fine.

In an attempt to resolve the issue permanently, added a local
insecurecall function which manipulates the secure state of the
environment directly and doesn't pass errors through to the global
error handler.