Rework error capturing for securecall-based tests
A new securecall_checked function sets a local error handler for the
duration of a single test which stashes the error away in an upvalue
that's checked upon return of the invoked function.

This should open back up routes for testing things involving securecall
and not needing to force the secure state of the environment which
isn't too reliable.