Rework CMake build setup
Added a few configuration toggles for fast math compiler options and
building with a C++ compiler for exception support, to closer match
the WoW environment.

Removed the lua-core object target and simply made the luac target
pull in the source list directly for building if liblua is shared,
since luac was having exported defines present in such a case which
was slightly irksome.