Add integer overflow errors on string.format calls
This fixes an inconsistency with string.format is called with an
integer format specifier like %d, %u, %x, or %o with a value that
is outside of the range of a signed long.

Note that in 9.1 this behaviour was changed for unsigned specifiers to
not error - this isn't replicated here yet as it'll probably need to
be implemented as a runtime configurable.