Added WoW-compatible bit library implementation
Note that this is largely just a quick and dirty first-attempt effort,
some verification has been done and things appear to function as they
do ingame.

The key difference between this built-in and automatically-loaded
bit library versus the community 'bitlib' project is primarily in two
areas; the function offering is different ('bit.mod' exists, but things
like 'bit.tobit' don't) and that values returned by most functions are
unsigned 32-bit integers, with the exception of 'bit.arshift' and
'bit.mod' which are signed.