Initial work on taint logging
This first pass extends the lua_getinfo/debug.getinfo APIs with a new
flag - "s" - which will return security related information about the
global VM.

From a C perspective, this flag will fill in the new taint field on
the lua_Debug structure with the global taint status of the VM. In Lua,
the debug.getinfo call will return a new field on the returned table
named "taintedby" which will be a string with the same content as the
"source" field on the lua_TaintInfo structure.

To know when the execution state has transitioned from either secure
to insecure or the other way around, a new "security" hook has been
added which can also be registered from Lua via the debug.sethook
API with flag "s". This will cause a user-defined function to execute
after the processing of any single instruction has completed which
caused the VM to change its overall security state.

At present, there's a few limitations to all of the above.

The first is that debug hooks are executed with the current security
state of the VM retained; so the "security" hook may further taint
execution if it's otherwise not careful. This will likely be fixed
with changes to the error handling logic to permit saving and restoring
a taint flag.

Secondly the taint information returned by the debug APIs is that of
the global VM state; it's not (yet) possible to query individual call
frames on the stack to know whether or not they were tainted prior to
entering their descendant frames.