Prefer using gettaint/settaint macros
This is mostly to make grepping for tweaks to taint fields a bit easier
instead of keeping some accesses to the taint field in naked form.

Added a new utility macro, propagatetaint, which takes a destination
and source parameter; if the source has taint applied then the taint
field on the destination will be replaced with that of the source.