Add xoroshiro based RNG for math.random
This implements a secondary less-predictable source of RNG that can
be used to differentiate the implementation of math.random vs fastrandom
so that one is less predictable than the other. Both sources are
seeded via the existing math.randomseed function, and the new source
has a fixed seed on startup similar to fastrandom.

This is a bit of a hack in some places - calling the math.random
function during shutdown of a Lua state might cause some issues, but
hopefully no one ever has a reason to do that.