Remove const qualifier from stored taint
As the API returns mutable taint info objects it's possible for any
statically allocated const taint info objects to incorrectly be
returned to callers, which could then go on to mutate the read-only
data segment.

These APIs were due for a cleanup and this is largely it - the API
surface and internals no longer consider taint info objects to be
constant. The forceinsecure function no longer places its own taint
object in the read-only data segment as well.

Additionally the lua_taintvalues function has been renamed to
lua_setstacktaint for consistency with other functions.