Lower C standard in CMakeLists to C11
CMake 3.16 doesn't know that C11 exists. The expectation was that
if it didn't know C17 existed then, because C_STANDARD_REQUIRED was
set to false, it'd pick the highest one that it does know.

This is not the case.

As C11 and C17 are equivalent minus some DRs and a version define, we'll
lower it just to accomodate CMake here and re-enable the
C_STANDARD_REQUIRED option to enforce C11 support.

This additionally reformats some parts of the CMakeLists.txt files.