Don't apply taint via setnilvalue
Given a situation where a table backed by a metatable which is proxying
all writes to another table, if an assignment occurs then the initial
table is resized during the search for the node prior to the metatable
being consulted. The table resize operation goes on to call setnilvalue
and creates a tainted nil value in the front table, which doesn't match
reference behavior.

This can be observed with the following snippet:

  forceinsecure()
  local a = {}
  local b = setmetatable({}, { __newindex = a })
  b.test = {}
  assert(not issecurevariable(a, "test"))
  assert(issecurevariable(b, "test"))

The behavior being seen was that the second assertion would fail as
the empty slot created during the search for `b.test` prior to its
assignment (through a metatable) would be tainted due to setnilvalue.

It is observable that nil values - including constants - are meant
to taint in insecure contexts. Loading the following insecurely will
demonstrate this:

  local b = {}
  b.test = nil
  assert(not issecurevariable(b, "test"))

Another concern is would be function calls that expect more return
values that are given. The luaD_poscall function would fill the
remainder of the "wanted" values with nils via setnilvalue, however an
observable in the reference is that this shouldn't be the case. The
following test case will exhibit this. Note that if the table isn't
preallocated with two hash nodes, the second assignment would cause
a rehash and - thus - trigger our problem.

  local test = { a = nil, b = nil }
  test.a, test.b = nop()
  assert(not issecurevariable(test, "a"))
  assert(not issecurevariable(test, "b"))

As a result it seems likely that the reference implementation doesn't
apply taint to nil values written through setnilvalue - which in the
stock source has no access to the Lua state anyway.

This changeset removes the state parameter to the setnilvalue macro
and NULLs the field when used. A special case has been added to the
creation of function prototypes to propagate taint from the environment
to registered constants - which doesn't seem too farfetched - and for
post-calls to additionally apply taint to the nils it pushes onto
the stack for additional returns.