Remove propagation of taint to non-object values
This may be an incorrect move and, if so, will be reverted later. Some
tests need adding to establish a few invariants and to shed some light
on how taint is supposed to function since there's an inherent issue
with how constants are treated which implies that it _may_ be that
taint is supposed to be applied at a per-instruction level rather than
through the lobject.h macros.