Backport nparams/isvararg debug info from Lua 5.2
These may be useful things to have around for external verification and
testing of code, so might as well have them.

This additionally moves the location of the "taint" field in the
lua_Debug structure since the location it had been inserted into
could pose a binary incompatibility if loaded.