Experimentally pass taint on in push* APIs
The C APIs that "create" or push values onto the stack will now
apply taint to those values, keeping in line with a change from
a previous commit to remove this being universally done throughout
the whole VM through the set* macros for primitive values.

Note that for object values like strings and closures this change
is included for their push functions but isn't needed, as those
macros weren't affected. This is just covered now in case that one
day changes.