Add taint forcing functions to auxilliary library
A few places have called for a static declaration of a taint info
object, so might as well make this part of the aux lib for reuse.

One small change this entails is that calling forceinsecure() will no
longer override any existing taint, which is the same as the ingame
environment.