Added method to embed Lua scripts into tests
This relocates the tests to a "tests" folder (renamed from "test")
and adds a way to embed Lua scripts as a precompiled blob into a
C source file as part of suite compilation.

The test utilities have been slightly reorganized to split things to
a new "common" file for functions, and we're now making use of the
TEST_INIT/TEST_FINI functionality of acutest to set up a global Lua
state for each test.