Rework taint handling macros
This overhauls the behaviour of taint and the existing Lua macros to
make setobj no longer apply taint to the Lua state when used, and
instead only be responsible for the transfer of taint from one
object to another.

This is _probably_ more accurate to reference as updating the taint
on the state - but not the other way around - feels somewhat incorrect.

The propagatetaint macro has been removed and a new set of luaO_*
macros for taint have been introduced to do the following:

  - Propagate taint from a Lua state to a value
  - Propagate taint from a value to a Lua state
  - Propagate taint from either a value to a state or vice versa
  - Copy taint from one object to another
  - Move taint from one object to another

This should offer a more fine-grained (albeit manual) approach to
handling taint that can be slowly peppered over the VM as more test
cases are added.

A couple of test cases have been added to verify behavior around
creating table structures with tainted values, and for reading secure
globals from a tainted context.