Only apply taint to stack in function precalls
This is _probably_ incorrect but will need more test cases to confirm,
however for now it fixes an issue whereby the stack slots once allocated
are uninitialized and could propagate a junk taint field back to the
state.

Realistically this should probably do a two-way application of taint
but capped from the base of the stack to the number of supplied
arguments.