Add luaT_loadfixture macro, remove loadfixtures
This allows tests to have fine-grained control and allows us to remove
the X-macro that I didn't much like to begin with. It also helps with
error reporting in tests since, as a macro, the printed messages will
use the file/line information for the test function itself rather than
the implementation in common.c.