Initial work on VM opcode taint propagation
As we're not overriding the setobj macros anymore taint needs peppering
in for a bunch of opcodes in luaV_execute. At the moment everything
up to OP_LEN has been briefly looked at and assessed, albeit with
tired post-vaccination eyes.

The general idea followed here is that:

  - Any reads from the B, C, Bx, and sBx registers will usually
    apply taint from the read value to the state.

  - Writes to the A register will almost always either apply taint
    from the state to the value, or do a two-way taint share between
    the value and state.

A few core exceptions exist; OP_LOADK for instance doesn't follow the
rules for the A register because of the secure constant anomaly, so that
instead will only apply taint from an insecurely loaded constant to the
state.

Another exception are the OP_SETGLOBAL, OP_SETTABLE, and OP_SETUPVAL
opcodes. These don't do any taint manipulation at all as it's assumed
that the value being written will already have had any relevant taint
applied at the point it was read or created.

Some of these are probably incorrect and need a lot of tests; it could
be that a few are too strict or that some are too lenient. OP_SELF
stands out as being a particularly odd one at present, and OP_CONCAT
though not implemented could be very tricky.