Make luaB_issecurevariable use lua_gettabletaint
This removes the need for the temporary taint stashing and simplifies
the implementation greatly. It'd probably be a good idea to apply this
same concept to securecall and hooksecurefunc implementations at some
point as well, since the idea we're going with is to have the core APIs
work behind the taint veil.