Make luaV_tonumber/luaV_tostring retain taint
The value post-conversion for luaV_tonumber would be stripped of any
taint; we now store the taint before the conversion and re-assign it
after.

For luaV_tostring this is less of an issue, however if for some reason
the converted string would be without taint despite having had some
originally it will now retain that taint.