Make C taint APIs more consistent
This changeset does the following:
  - Renames lua_gettaint to lua_getthreadtaint
  - Renames lua_settaint to lua_setthreadtaint
  - Adds lua_settabletaint and lua_setfieldtaint
  - Turns the luaL_issecure* macros to static inline functions.
  - Adds a general purpose luaV_taintbarrier macro

The gettaint/settaint functions were ambiguously named in that they
didn't explicitly make it clear that they worked on the thread object.

The luaV_taintbarrier macro should simplify a few cases where taint is
"stored" ahead of a potentially-tainting operation. A few places need
this sprinkled in, but it's currently used by the gettabletaint and
settabletaint implementations which previously lacked this.