Add debug APIs for taint CONTROL
This changeset expands the debug APIs with additional functions for
querying taint at a more granular level and for explicitly tainting
things with custom source names.

The following APIs have been added:

  - debug.issecureclosure(func)
  - debug.issecurefield([table,] key)
  - debug.issecurelocal([thread,] level, local)
  - debug.issecurethread([thread])
  - debug.issecureupvalue(func, up)
  - debug.setclosuretaint(func, source)
  - debug.setfieldtaint([table,] key, source)
  - debug.setlocaltaint([thread,] level, local, source)
  - debug.setthreadtaint([thread,] source)
  - debug.setupvaluetaint(func, up, source)

In each case above `source` may be nil to clear taint, or a string
name to taint a value.