Mark closures tainted if prototype is insecure
This is largely just an assumption and would probably be impossible to
actually test, as secure code would need to somehow run an insecure
function securely to generate a closure with, but regardless if a
Lua closure is created from a prototype defined insecurely then we'll
propagate taint to it if it wasn't already assigned by the thread.

Additionally this fixes an issue where the taint field on function
prototypes was left uninitialized.