Re-add state parameter to set[bnp]value macros
This causes booleans, numbers, and light pointers assigned to TValues
to inherit taint from the thread immediately upon assignment. This was
the case a while ago but was removed in favo of a more manual approach
for taint.

It's more likely that this is _probably_ how it's implemented due to
the number of changes required to the VM instruction loop for taint
to propagate otherwise for these values, so we'll assume this for now
and revert if later tests prove wrong.