Swap to FetchContent for dependency management
Submodules are nice but they're a bit annoying if you forget to
recursively grab them on initial clone. FetchContent on the other
hand is much nicer as it statically encodes our dependencies inside
the build system.

As neither of our dependencies have suitable CMake integrations as-is
this does require a bit of manual target generation, but it works out
in the end.