Ensure geterrorhandler returns correct functions
When calling seterrorhandler from scripts a proxy handler is installed
to handle the "UNKNOWN ERROR" case in the reference environment, which
makes geterrorhandler always return the proxy closure rather than the
closure supplied to seterrorhandler.

This corrects that and ensures the original closure is always returned,
such that the proxy is effectively invisible from scripts.