Add coroutine.bind() and coroutine.call()
These functions will allow you to call a given function with arguments
on a specified thread.

This bypasses the normal resume/yield machinery - so it's really only
intended for use cases where you need to call from a separate thread
back into the main thread. The use case here is to permit wowless to
effectively run its code on a "shadow" thread which - from the
perspective of addon code - only consumes a single level in the main
thread stack.