Implement script profiling APIs
This finishes the bulk of the work for the performance profiling APIs
allowing stats to be collected on the execution time of individual
functions.

By default, profiling is disabled. The state of CPU profiling can be
toggled at-will through the lua_setprofilingenabled() API. Enabling
profiling will enable accounting for all functions - even those created
before profiling was enabled.

Note that enabling profiling will naturally incur a heavy performance
penalty for any code that is particularly function-call heavy. Stats
are committed at points that appear to be consistent with the reference
implementation, so there are some quirks that might be noticed.