Remove unused security hook
This was disabled and removed with The Great Taint Rewrite as there
were some issues that led to its behavior being undesirable; notably
when taint would transition in the middle of C code *far* away from
the actual interpreted instruction.

Longer-term it'd be nice to bring over a taint event buffer that can
be queried on-demand.