Add elune_target_public_headers function
The handling of public headers is a bit ugly and may change in the
future if the CMake version is bumped to 3.23 to take advantage of
the added FILE_SET support.

To possibly accomodate this nicely in the future - and allow for other
public headers in side-targets to be exposed - this function has been
added to ensure consistent handling of public headers for installation.