Backport Lua 5.4 standalone interpreter
This is mostly a cleanup exercise as the sources for the 5.4 interpreter
binary are a little easier to manage with respect to arguments.

Functionally, this brings in a few changes:

  - The '-L' flag now disables the Lua standard library.
  - The '-w' flag now imports the WoW standard library.
  - The '-T' flag has been added and disables taint propagation if set.
  - The '-E' flag from newer versions is respected by the interpreter.

One change of note is that the global error handler is no longer
invoked for errors raised by the REPL. It will - however - still be
triggered for errors raised by calls to 'securecall' and any other
C APIs which use the global error handler.