Fix extern decls and attributes for CXX
Prefer using '__has_cpp_attribute' when available; unfortunately with
GCC at least it seems that __has_cpp_attribute is defined even in C
mode so it needs an additional gate based on '__cplusplus' being
defined.

For linkage defines, we'll need to use 'extern "C++"' explicitly to
avoid errors in some configurations with the positioning of the
noreturn attribute.