Add lua_objname extension API
This function will attempt to call an '__name' metamethod on a given
value, returning its value if present or nil if not.

The intended use case here is to provide a hook for any API aspects
that require a name of an object - such as error dumps involving frame
objects.