Resolved correctness issues with debuglocals
The initial implementation didn't ensure that stack operations were
balanced between each call to a buffer-appending API. While this worked
in limited testing, it probably would have led to some odd behaviors
in some cases.

The new implementation _should_ ensure that the stack is correctly
balanced between append operations. For table dumping this is a bit
trickier, and is handled by having a new sub-buffer be created with
a temporary scratch space for the key to be held in during table
traversal.