Add name replacement support to error handler
The reference client has a quirk whereby if the "source" associated with
a stack frame that triggers an error contains a "*:" token then it will
replace the "*" with the name of any frame-like object that is stored
in the first local value of the stack frame.

This is largely only used for script handlers defined in XML, such
as the following case where the error message will contain the
name of the frame or "<unnamed>" if none is available.

	<Frame name="Banana" setAllPoints="true">
		<Scripts>
			<OnMouseDown>a = a + 1</OnMouseDown>
		</Scripts>
	</Frame>

If you inspect the stack trace you'll see that the source associated
with the stack frame is a literal "*", which gets printed out as
"*:OnMouseDown" as "OnMouseDown" is a method name.

This can be replicated in pure Lua through loadstring; supply a source
name that contains "*:" right at the start and it'll replace the "*"
with the name of the frame stored in the first local (or parameter)
of the executed function:

    /run loadstring("local _ = ...; a = a + 1", "*:MyScript")(UIParent)

This appears to be a quirk of the error handler and nothing else, as
this replacement does not occur in the stack trace body generated by
the debugstack API.

As such it has been implemented as accurate to reference as is possible
at this stage. For object names it'll use the same "__name" metamethod
on a userdata, defaulting to "<unnamed>" if a non-stringable value is
returned. If no metamethod exists, it will assume it's not a frame-like
object and do no replacements.