Add 'n' parameter to lua_cleartaint
The expectation of what lua_cleartaint should actually be responsible
for clearing has never been too clear with respect to stack variables,
so to make it slightly easier a new 'n' parameter has been added to
qualify how much of the top of the stack you want cleared.

Specifying zero will result in the previous behavior of the API wherein
it'll clear no stack value taint, and lua_gettop(L) can be used to clear
the whole stack of taint.