Add fast math flag option
The fast math flag is required for accuracy against a reference
client environment, so having this be part of the CMake presets
mechanism is a bit misleading and causes trouble elsewhere for
consumption of the project.

This flag (and potentially others to follow) will now be handled
directly by the CMakeLists file for the project, guarded behind
options that can be toggled at-will.