Disallow NaN values as table keys
This change was made in patch 10.2.0 and should be universally applied
now, though perhaps not to Wrath Classic.

The NaN detection logic was fixed and so the reference client now
correctly disallows NaN values to be used as table keys; this is
the same as a stock build of Lua with math optimizations disabled.