Fix incorrect handling of taint in hooksecurefunc
The original test case wasn't particularly exhaustive and didn't check
the current execution stack taint as well as it should have, leading
to an inaccuracy in the handling of cases whereby an insecure global
function is hooked by secure code.

This should almost never happen in practice, but with the added
assertions it was passing in the reference client but failing in
our implementation.