Always assign new keys into tables securely
This is actually a rather surprising correctness bug; the insertion of
new keys into tables has the effect of always cleaning the key of all
taint.

This can be observed with a simple constant assignment test, followed by
attempting to iterate over the contents of the table with next or pairs.
Previously, the iteration would taint execution upon retrieving the key,
however the actual reference client behavior does not align with this
and execution there remains secure.

The likely cause of this difference is just down to the original Lua
source, which handles assignment of objects in the newkey function a
little bit more manually than in other places.